# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для системы осмотра жилых помещений ЖКХ.\nСистема предназначена для координации работы инспекторов,\nуправления заданиями на осмотр зданий и формирования отчётности.",
        "title": "JKH Inspection API",
        "contact": {
            "name": "Техническая поддержка",
            "email": "support@jkh.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/checklists": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех чек-листов (без детализации элементов)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Чек-листы"
                ],
                "summary": "Получить список чек-листов",
                "responses": {
                    "200": {
                        "description": "Список чек-листов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ChecklistResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создание нового чек-листа для осмотра зданий",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Чек-листы"
                ],
                "summary": "Создать чек-лист",
                "parameters": [
                    {
                        "description": "Данные чек-листа",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateChecklistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Чек-лист успешно создан",
                        "schema": {
                            "$ref": "#/definitions/models.ChecklistResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Чек-лист с таким названием уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/checklists/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает детальную информацию о чек-листе включая список элементов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Чек-листы"
                ],
                "summary": "Получить чек-лист по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чек-листа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные чек-листа с элементами",
                        "schema": {
                            "$ref": "#/definitions/models.ChecklistDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Чек-лист не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновление данных чек-листа (название, тип осмотра)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Чек-листы"
                ],
                "summary": "Обновить чек-лист",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чек-листа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateChecklistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные чек-листа",
                        "schema": {
                            "$ref": "#/definitions/models.ChecklistResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Чек-лист не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Название чек-листа уже занято",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление чек-листа из системы",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Чек-листы"
                ],
                "summary": "Удалить чек-лист",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чек-листа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Чек-лист успешно удален"
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Чек-лист не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Чек-лист используется в заданиях",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/checklists/{id}/elements": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавление элемента из справочника в чек-лист",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Чек-листы"
                ],
                "summary": "Добавить элемент в чек-лист",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чек-листа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ID элемента и порядок",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddElementToChecklistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Элемент успешно добавлен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или элемент не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Чек-лист не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Элемент уже добавлен в чек-лист",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddElementToChecklistRequest": {
            "type": "object",
            "required": [
                "element_id"
            ],
            "properties": {
                "element_id": {
                    "description": "ID элемента из справочника ElementCatalog.",
                    "type": "integer",
                    "minimum": 1
                },
                "order_index": {
                    "description": "Порядок проверки элемента в чек-листе (опционально).\nЕсли не указан, элемент добавляется в конец списка.",
                    "type": "integer"
                }
            }
        },
...
        "models.ChecklistDetailResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "elements": {
                    "description": "Список элементов в чек-листе",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChecklistElementDetail"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "inspection_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ChecklistElementDetail": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Категория элемента",
                    "type": "string"
                },
                "element_id": {
                    "description": "ID элемента из ElementCatalog",
                    "type": "integer"
                },
                "element_name": {
                    "description": "Название элемента (например, \"Кровля\")",
                    "type": "string"
                },
                "order_index": {
                    "description": "Порядок проверки (1, 2, 3...)",
                    "type": "integer"
                }
            }
        },
...
        "models.ChecklistResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "ISO 8601 формат",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inspection_type": {
                    "description": "spring/winter/partial",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateChecklistRequest": {
            "type": "object",
            "required": [
                "inspection_type",
                "title"
            ],
            "properties": {
                "description": {
                    "description": "Описание чек-листа (опционально).",
                    "type": "string"
                },
                "inspection_type": {
                    "description": "Тип осмотра: \"spring\" (весенний), \"winter\" (зимний), \"partial\" (частичный).\nЗначение по умолчанию в БД: \"partial\".",
                    "type": "string",
                    "enum": [
                        "spring",
                        "winter",
                        "partial"
                    ]
                },
                "title": {
                    "description": "Название чек-листа (например, \"Весенний осмотр многоквартирных домов\").\nПоле обязательно и уникально в БД.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите JWT токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}